requirements.txt
streamlit
torch==2.0.1
torchaudio
pyannote.audio==2.1.1
numpy
scikit-learn
soundfile
sounddevice
av